♻️ separate template component for reuse
